공지사항 로그인&권한 체크 로직, 테스트 추가

+++ src/main/java/kr/co/pinup/notices/controller/NoticeApiController.java
@@ -1,6 +1,11 @@
 package kr.co.pinup.notices.controller;
 
 import jakarta.validation.Valid;
+import kr.co.pinup.custom.loginMember.LoginMember;
+import kr.co.pinup.exception.common.ForbiddenException;
+import kr.co.pinup.exception.common.UnauthorizedException;
+import kr.co.pinup.members.model.dto.MemberInfo;
+import kr.co.pinup.members.model.enums.MemberRole;
 import kr.co.pinup.notices.service.NoticeService;
 import kr.co.pinup.notices.model.dto.NoticeCreateRequest;
 import kr.co.pinup.notices.model.dto.NoticeUpdateRequest;
@@ -33,23 +38,47 @@ public class NoticeApiController {
     }
 
     @PostMapping
-    public ResponseEntity<Void> save(@RequestBody @Valid NoticeCreateRequest request) {
-        noticeService.save(request);
+    public ResponseEntity<Void> save(@LoginMember MemberInfo memberInfo,
+                                     @RequestBody @Valid NoticeCreateRequest request) {
+        ensureAuthenticated(memberInfo);
+        ensureAdminRole(memberInfo);
+
+        noticeService.save(memberInfo, request);
 
         return ResponseEntity.status(CREATED).build();
     }
 
     @PutMapping("/{noticeId}")
-    public ResponseEntity<Void> update(@PathVariable Long noticeId, @RequestBody @Valid NoticeUpdateRequest request) {
+    public ResponseEntity<Void> update(@LoginMember MemberInfo memberInfo,
+                                       @PathVariable Long noticeId,
+                                       @RequestBody @Valid NoticeUpdateRequest request) {
+        ensureAuthenticated(memberInfo);
+        ensureAdminRole(memberInfo);
+
         noticeService.update(noticeId, request);
 
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{noticeId}")
-    public ResponseEntity<Void> delete(@PathVariable Long noticeId) {
+    public ResponseEntity<Void> delete(@LoginMember MemberInfo memberInfo, @PathVariable Long noticeId) {
+        ensureAuthenticated(memberInfo);
+        ensureAdminRole(memberInfo);
+
         noticeService.remove(noticeId);
 
         return ResponseEntity.noContent().build();
+    }
+
+    private void ensureAuthenticated(MemberInfo memberInfo) {
+        if (memberInfo == null) {
+            throw new UnauthorizedException("인증이 필요합니다.");
+        }
+    }
+
+    private void ensureAdminRole(MemberInfo memberInfo) {
+        if (MemberRole.ROLE_ADMIN != memberInfo.getRole()) {
+            throw new ForbiddenException("액세스할 수 있는 권한이 없습니다.");
+        }
     }
 }
